refactor: move buffer module to rails.llm directory
